web_ui_framework/requirements.txt
selenium>=4.0.0
pytest>=7.0.0
pyyaml>=6.0
allure-pytest>=2.0.0
pytest-html>=3.0.0
